+++ nidl/transforms/volume/augmentation/intensity/random_noise.py
@@ -14,7 +14,7 @@ from typing import Union
 import numpy as np
 import torch
 
-from .....transforms import TypeTransformInput, VolumeTransform
+from ....transforms import TypeTransformInput, VolumeTransform
 
 
 class RandomGaussianNoise(VolumeTransform):
