+++ nidl/transforms/volume/preprocessing/spatial/crop_or_pad.py
@@ -12,7 +12,7 @@ from typing import Union
 import numpy as np
 import torch
 
-from .....transforms import TypeTransformInput, VolumeTransform
+from ....transforms import TypeTransformInput, VolumeTransform
 
 
 class CropOrPad(VolumeTransform):
